requirements.txt
python-telegram-bot>=20.0
google-generativeai>=0.3.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
aiohttp>=3.8.5